+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 Unreleased
 
+- `confirm()` and `prompt()` now properly strip ANSI codes from
+  the prompt text when color is disabled. {issue}`3572` {pr}`3599`
 - Supported versions of Windows enable ANSI terminal styles by default.
   Colorama is no longer a dependency and is not used. {issue}`2986` {pr}`3505`
 - {class}`Argument` accepts a `help` parameter, and help output includes